Bypass networking when virtualized

+++ kickstart/scripts/bridge-deploy.sh
@@ -1,31 +1,34 @@
 #!/bin/bash
 
 deploy_bridge_ubuntu() {
-	# enable port forwarding
-	expand etc/sysctl.d/99-forwarding.conf /etc/sysctl.d/99-forwarding.conf
+  local v="`virt-what 2>/dev/null`"
+  if [ $? = 0 ] && [ -z "$v" ]; then
+		# enable port forwarding
+		expand etc/sysctl.d/99-forwarding.conf /etc/sysctl.d/99-forwarding.conf
 
-	# load sysctl settings
-	service procps start
+		# load sysctl settings
+		service procps start
 
-	# configure interface hook scripts
-	expand etc/enable-port-forwarding /etc/network/if-up.d/enable_port_forwarding
-	expand etc/disable-port-forwarding /etc/network/if-down.d/disable_port_forwarding
-	chmod +x /etc/network/if-up.d/enable_port_forwarding
-	chmod +x /etc/network/if-down.d/disable_port_forwarding
+		# configure interface hook scripts
+		expand etc/enable-port-forwarding /etc/network/if-up.d/enable_port_forwarding
+		expand etc/disable-port-forwarding /etc/network/if-down.d/disable_port_forwarding
+		chmod +x /etc/network/if-up.d/enable_port_forwarding
+		chmod +x /etc/network/if-down.d/disable_port_forwarding
 
-	IFACE=$posm_wan_netif /etc/network/if-up.d/enable_port_forwarding
+		IFACE=$posm_wan_netif /etc/network/if-up.d/enable_port_forwarding
 
-	# disable DNS wildcarding
+		# disable DNS wildcarding
 
-	rm /etc/dnsmasq.d/99-captive.conf
+		rm /etc/dnsmasq.d/99-captive.conf
 
-	service dnsmasq restart
+		service dnsmasq restart
 
-	# disable Nginx captive portal
+		# disable Nginx captive portal
 
-	rm -f /etc/nginx/sites-enabled/captive
+		rm -f /etc/nginx/sites-enabled/captive
 
-	service nginx restart
+		service nginx restart
+	fi
 }
 
 deploy bridge
